+++ modules/libraries/core/build.gradle
@@ -1,11 +1,3 @@
-/*
- * This file was generated by the Gradle 'init' task.
- *
- * This generated file contains a sample Java library project to get you started.
- * For more details take a look at the 'Building Java & JVM projects' chapter in the Gradle
- * User Manual available at https://docs.gradle.org/8.0-rc-1/userguide/building_java_projects.html
- */
-
 description = 'jMe3GL2 - core'
 
 dependencies {
