+++ modules/libraries/dyn4j/build.gradle
@@ -1,0 +1,8 @@
+description = 'jMe3GL2 - dyn4j'
+
+dependencies {
+    implementation project(':modules:libraries:jMe3GL2-core')
+
+    implementation 'org.jmonkeyengine:jme3-core:' + jmonkeyengineVersion
+    implementation 'org.dyn4j:dyn4j:' + dyn4jVersion
+}
