+++ modules/libraries/dyn4j/src/main/java/org/je3gl/physics/ThreadingType.java
