+++ modules/libraries/dyn4j/src/main/java/org/je3gl/physics/debug/Dyn4jDebugColor.java
