+++ modules/libraries/dyn4j/src/main/java/org/je3gl/scene/tile/SpritesheetPhysicsAdapter.java
@@ -37,66 +37,62 @@ import org.je3gl.physics.PhysicsSpace;
 import org.je3gl.physics.control.PhysicsBody2D;
 
 /**
- * Interface <code>SpritesheetPhysics</code> is in charge of managing the changes
- * of a {@link TileMap}.
- * 
- * <p>
- * With this interface we can know and manage every change made by the geometries
- * of the parent node.
- * </p>
- * 
- * <p>
- * Note that the {@link TileMap} inherits from <code>GeometryGroupNode</code> so
- * most of the events of this interface correspond to the abstract implementation
- * of the methods of the parent of the class {@link TileMap}.
- * </p>
- * 
+ * Convenience class to implement the interface {@link SpritesheetPhysicsAdapter}.
  * @author wil
  * @version 1.0.5
  * @since 2.0.0
+ * 
+ * @deprecated Choose to use <a href="https://github.com/jmecn/TMXLoader">TMXLoader</a> for 2D tilemaps
  */
-public interface SpritesheetPhysics {
-    
-    /**
-     * Method in charge of establishing the physical space of the rigid bodies
-     * of the tile map.
-     * @param physicsSpace physical space
+@Deprecated(since = "3.1.0")
+public abstract class SpritesheetPhysicsAdapter implements SpritesheetPhysics {
+
+    /*
+     * (non-Javadoc)
+     * @see SpritesheetPhysics#setPhysicsSpace(jMe3GL2.physics.PhysicsSpace) 
      */
-    public void setPhysicsSpace(PhysicsSpace<PhysicsBody2D> physicsSpace);
-    
-    /**
-     * It is activated when a geometry is removed from the parent node.
-     * @param geom geometry eliminated.
+    @Override
+    public void setPhysicsSpace(PhysicsSpace<PhysicsBody2D> physicsSpace) { }
+
+    /*
+     * (non-Javadoc)
+     * @see SpritesheetPhysics#onDetachTile(com.jme3.scene.Geometry) 
      */
-    public void onDetachTile(Geometry geom);
-    
-    /**
-     * It is activated when a new geometry is added to the parent node.
-     * @param geom new geometry
+    @Override
+    public void onDetachTile(Geometry geom) { }
+
+    /*
+     * (non-Javadoc)
+     * @see SpritesheetPhysics#onAttachTile(com.jme3.scene.Geometry) 
      */
-    public void onAttachTile(Geometry geom);
-    
-    /**
-     * The geometry is no longer associated with the parent node.
-     * @param geom geometry
+    @Override
+    public void onAttachTile(Geometry geom) { }
+
+    /*
+     * (non-Javadoc)
+     * @see SpritesheetPhysics#onTileUnassociated(com.jme3.scene.Geometry) 
      */
-    public void onTileUnassociated(Geometry geom);
-    
-    /**
-     * It is activated when a change is applied to a geometry.
-     * @param geom geometry
+    @Override
+    public void onTileUnassociated(Geometry geom) { }
+
+    /*
+     * (non-Javadoc)
+     * @see SpritesheetPhysics#onTransformChange(com.jme3.scene.Geometry) 
      */
-    public void onTransformChange(Geometry geom);
-    
-    /**
-     * It is activated when a change in the material of a geometry is applied.
-     * @param geom geometry
+    @Override
+    public void onTransformChange(Geometry geom) { }
+
+    /*
+     * (non-Javadoc)
+     * @see SpritesheetPhysics#onMaterialChange(com.jme3.scene.Geometry) 
      */
-    public void onMaterialChange(Geometry geom);
-    
-    /**
-     * It is activated when the mesh or level of a geometry is changed.
-     * @param geom geometry
+    @Override
+    public void onMaterialChange(Geometry geom) { }
+
+    /*
+     * (non-Javadoc)
+     * @see SpritesheetPhysics#onMeshChange(com.jme3.scene.Geometry) 
      */
-    public void onMeshChange(Geometry geom);
+    @Override
+    public void onMeshChange(Geometry geom) { }    
 }
