+++ modules/libraries/dyn4j/src/main/java/org/je3gl/scene/tile/Tile.java
@@ -46,7 +46,10 @@ import java.io.IOException;
  * @author wil
  * @version 1.0.5
  * @since 2.0.0
+ * 
+ * @deprecated Choose to use <a href="https://github.com/jmecn/TMXLoader">TMXLoader</a> for 2D tilemaps
  */
+@Deprecated(since = "3.1.0")
 public class Tile implements Savable, Cloneable {
     
     /** Properties of this object. */
