+++ modules/libraries/dyn4j/src/main/java/org/je3gl/scene/tile/TileMapGroupNode.java
@@ -42,7 +42,10 @@ import com.jme3.scene.Node;
  * @version 1.0.5
  * @since 2.0.0
  * @param <T> data type
+ * 
+ * @deprecated Choose to use <a href="https://github.com/jmecn/TMXLoader">TMXLoader</a> for 2D tilemaps
  */
+@Deprecated(since = "3.1.0")
 public class TileMapGroupNode<T extends TileMap> extends Node {
     /* CODE */
 }
