+++ modules/samples/build.gradle
@@ -32,6 +32,7 @@ dependencies {
     implementation 'org.dyn4j:dyn4j:' + dyn4jVersion
 
     implementation project(':modules:libraries:jMe3GL2-core')
+    implementation project(':modules:libraries:jMe3GL2-dyn4j')
     implementation project(':modules:libraries:jMe3GL2-jawt')
     implementation project(':modules:libraries:jMe3GL2-plugins')
 }
